Fixing Issue #27
extra_context is being passed a RequestContext object so when the update
method of the local context object is called it assumes it is being
passed another dict and since it's not it triggers an error.

As extra_context is defaulted to {} and both dict and RequestContext
objects have update methods this is a safe fix that returns the form to
a working state with django & django-cms.